voxgrambot/requirements.txt
BTrees==4.3.1
certifi==2016.8.31
ExtensionClass==4.1.2
future==0.15.2
persistent==4.2.1
python-json-logger==0.1.5
python-telegram-bot==5.0.0
requests==2.11.1
six==1.10.0
transaction==1.6.1
urllib3==1.16
zc.lockfile==1.2.1
ZConfig==3.1.0
zdaemon==4.1.0
ZEO==5.0.0b0
ZODB==4.4.3
ZODB3==3.11.0
zodbpickle==0.6.0
zope.interface==4.3.1
